Rename utils.h to avoid collision with library.
